:sparkles: Created Layout render
Created Layout render system, added default layout and moved all animation logic to this layout from App.vue.
